:wheelchair: Améliore l'accessibilité du composant de recherche de commune"
